refactor login actions
