+++ frontend/src/Navbar/Navbar.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import { StyleSheet, css } from 'aphrodite';
 import { connect } from 'react-redux';
-import {logout} from '../Redux/actions';
+import {logout} from '../Redux/loginActions';
 
 class navbar extends Component {
   static propTypes = {
